Renamed StyledTitle  component
